+++ backend/src/middleware/admin.middleware.js
@@ -2,8 +2,9 @@ import isAdmin from '../helpers/isAdmin.js';
 
 export default async function (req, res, next) {
     try {
-        if (!isAdmin(req.user)) throw new Error('Invalid permissions');
-
+        if (!isAdmin(req.user)) {
+            throw new Error('Invalid permissions');
+        }
         next();
     } catch (err) {
         return res.status(403).json({ message: err.message });